changes + agent-payload version

+++ pkg/security/ebpf/c/include/hooks/network/bind.h
@@ -75,7 +75,7 @@ int __attribute__((always_inline)) sys_bind_ret_impl(void *ctx, int retval, enum
         bind_key.addr[0] = event->addr[0];
         bind_key.addr[1] = event->addr[1];
 
-        u32 bind_cookie = 0;
+        u64 bind_cookie = 0;
         u32 bind_refresh_needed = 0;
         if (approve_bind_sample(&bind_key, &bind_cookie, &bind_refresh_needed) == SAMPLED) {
             event->event.flags |= EVENT_FLAGS_ACTIVITY_DUMP_SAMPLE | EVENT_FLAGS_SAVED_BY_AD;
