+++ pkg/security/secl/model/model_unix.go
@@ -666,7 +666,7 @@ type OpenEvent struct {
 	Flags uint32    `field:"flags"`                 // SECLDoc[flags] Definition:`Flags used when opening the file` Constants:`Open flags`
 	Mode  uint32    `field:"file.destination.mode"` // SECLDoc[file.destination.mode] Definition:`Mode of the created file` Constants:`File mode constants`
 
-	SampleCookie uint32 `field:"-"`
+	SampleCookie uint64 `field:"-"`
 
 	// Syscall context aliases
 	SyscallPath  string `field:"syscall.path,ref:open.syscall.str1"`  // SECLDoc[syscall.path] Definition:`Path argument of the syscall`
@@ -897,7 +897,7 @@ type BindEvent struct {
 	Addr         IPPortContext `field:"addr"`        // Bound address
 	AddrFamily   uint16        `field:"addr.family"` // SECLDoc[addr.family] Definition:`Address family`
 	Protocol     uint16        `field:"protocol"`    // SECLDoc[protocol] Definition:`Socket Protocol`
-	SampleCookie uint32        `field:"-"`
+	SampleCookie uint64        `field:"-"`
 }
 
 // ConnectEvent represents a connect event
@@ -908,13 +908,13 @@ type ConnectEvent struct {
 	Hostnames    []string      `field:"addr.hostname,handler:ResolveConnectHostnames,opts:skip_ad|root_domain|length"` // SECLDoc[addr.hostname] Definition:`Address hostname (if available)`
 	AddrFamily   uint16        `field:"addr.family"`                                                                   // SECLDoc[addr.family] Definition:`Address family`
 	Protocol     uint16        `field:"protocol"`                                                                      // SECLDoc[protocol] Definition:`Socket Protocol`
-	SampleCookie uint32        `field:"-"`
+	SampleCookie uint64        `field:"-"`
 }
 
 // SampleRefreshEvent is a lightweight internal event sent when a dedup map
 // detects a duplicate and wants to refresh the cookie timestamp in userspace.
 type SampleRefreshEvent struct {
-	Cookie uint32
+	Cookie uint64
 }
 
 // AcceptEvent represents an accept event
@@ -956,7 +956,7 @@ type SyscallsEvent struct {
 	EventReason  SyscallDriftEventReason
 	Syscalls     []Syscall // 64 * 8 = 512 > 450, bytes should be enough to hold all 450 syscalls
 	SyscallID    uint32    // populated iff EventReason == SampleReason
-	SampleCookie uint32    // populated iff EventReason == SampleReason
+	SampleCookie uint64    // populated iff EventReason == SampleReason
 }
 
 // PathKey identifies an entry in the dentry cache
