+++ pkg/security/security_profile/profile_manager.go
@@ -64,5 +64,5 @@ type ProfileManager interface {
 
 	// HandleSampleRefresh updates the LastSeen timestamp for the activity tree node
 	// associated with the given kernel dedup cookie
-	HandleSampleRefresh(cookie uint32)
+	HandleSampleRefresh(cookie uint64)
 }
